🦄 refactor: add exception in create PTO api

+++ src/modules/pto/commands/delete-pto-detail/delete-pto-datail.service.ts
@@ -1,14 +1,19 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs'
-import { PtoNotFoundException } from '../../domain/pto.error'
+import { PaidPTOException, PtoNotFoundException } from '../../domain/pto.error'
 import { DeletePtoDetailCommand } from './delete-pto-detail.command'
 import { PtoRepository } from '../../database/pto.repository'
 
 @CommandHandler(DeletePtoDetailCommand)
-export class DeletePtoService implements ICommandHandler {
+export class DeletePtoDetailService implements ICommandHandler {
   constructor(private readonly ptoRepository: PtoRepository) {}
   async execute(command: DeletePtoDetailCommand): Promise<void> {
-    const entity = await this.ptoRepository.findOne(command.ptoDetailId)
+    const entity = await this.ptoRepository.findOnePtoDetail(command.ptoDetailId)
     if (!entity) throw new PtoNotFoundException()
+
+    if (entity.getProps().isPaid) {
+      throw new PaidPTOException()
+    }
+
     await this.ptoRepository.delete(command.ptoDetailId)
   }
 }
